Zack's forest lab
This was a fun lab, I especially enjoyed learning how to make fog. I was unable to block out the background lighting, however, so I'll have to look into that eventually. I think the sun might be a little off in direction, but for some reason I couldn't adjust its position so I'll have to test that on other levels too to see if that problem continues.